move common code to evpkey.go
